Changed flow: cmd/scalecloud.de-api is the entry point.
The goal is to check if all connections and files are present before starting api endpoints.
The entry point starts api.go. This starts init from all the others (mongo, stripe, firebase).
When they have all been able to connect to their endpoints, api.go starts initializing the endpoints. Each endpoint will call scalecloud.de-api, which acts as a manager and has logic as to what needs to be done (logic) to get the desired result.