api: Added router group "Dashboard" which can be accessed only after authentication.
api: Added router handler AuthRequired to check if basicAuth was passed with UID and token. If not valid, 401 Unauthorized will be issued.
firebase: VerifyIDToken has been made public and now does not require an initialized firebase app as it fetches it itself.
scalecloud.de-api: As the main manager, it now only redirects IsAuthenticated so that firebase can be accessed to verify UID and token.