Started to implemented stripe api.api
Changed a lot of log messages to debug.
It is not possible to read the api key and use it to make requests to stripe.
Removed placeholder array of subscriptions for subscription overview.
Subscriptions for the subscription overview are now going to be read from stripe api
Moved methods to manager classes to better differentiate where to find what methods.
Changed named of models to make it more obvious that they hold a struct.