Implemented mono manager that provides create, update, delete, find without types. Each call to the methods gets a client and the corresponding collection and closes the database connection again.
mongo-stripe-manager uses the user-model to implement create, updating, delete and find more easily. In it are mainly methods for use with the database stripe and collection users.
mongo-stripe-manager also has an init method to check at startup if a connection to database users is possible. It also expects that at least one user exists. If not, a warning is issued.
Added user model to simplify the handling of users
go get -u ./...