A token is required for some routes. Verification of the token and whether it has been revoked is now done only once, which should improve performance by about 400 ms.
If a handler for a route requires the UID or email of the token, also known as TokenDetails, only a verifyIDToken is now performed to decode the token. The token will not be verified a second time if it has been revoked.
The method to retrieve the Bearer token has been moved to the Firebase Manager and to the method to retrieve TokenDetails. This has reduced duplicate code.