I have started to implement that it is possible to manage seats. First implementation is reading saved seats from the Mongodb via API.
New endpoint defined.
Moved all MongoDB databases and collections to one place. Added additional startup check if databases and collections exist. This facilitates the transition from the test phase to production without forgetting a collection. Also, a collection should not be created when something goes wrong.
Additional check at startup whether a unique index exists for a seat collection. If not, it will be created. This is helpful when moving everything to a new MongoDB.
Currently, only findOne was needed to fetch MongoDB documents. Now that I am looking for many seats, the new implementation will be findDocuments, where the cursor adds all to an typed slice.
I have prepared the mongo-seat-manager for create, find and delete, so more endpoints can be added soon.
Finished the implementation of the first endpoint for GetSubscriptionListSeat.
Prepared some new structs for later endpoints.